--- src/raspi_music_button/__main__.py
@@ -1,4 +1,0 @@
-from raspi_music_button.entrypoints.cli import cli
-
-if __name__ == "__main__":
-    cli()
