+++ src/raspi_music_button/entrypoints/cli/__init__.py
@@ -1,0 +1,1 @@
+
